+++ tests/input/enums.gd
@@ -1,6 +1,5 @@
 # Enum spacing and wrapping
-# Single line enum get spaces between members, multi-line enums get 1 indent and
-# trailing commas.
+# Enums are always vertical, with one indent and trailing commas.
 enum {A,B,C}
 
 
@@ -23,3 +22,5 @@ func foo():
 
 enum Test {
 	A, B }
+
+enum Single { ONLY }